realtime-livekit-nova2/pyproject.toml
[project]
name = "livekit-custom-nova2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-sdk-bedrock-runtime>=0.3.0",
    "aws-sdk-signers>=0.1.0",
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-aws>=1.4.2",
    "python-dotenv>=1.2.1",
]
